# BarCrawler
Given a location and a search term, we will make you a customizable tour of the neighborhood to find the hottest spots.

## Contributors

[Michael Garod](http://github.com/mgarod) DevOps, Full Stack Engineer

![alt text](https://github.com/mgarod/mgarod.github.io/blob/master/images/self.png)

Michael has been working as a Computer science and tutor at Hunter College for 2 years, and has recently completed a summer with Barclays as a Technology Analyst. Before that, he was a research assistant for Professor Carsten Kessler at Hunter College studying Wikipedia's role in geography theory. Michael has a passion for databases and big data technologies, cloud computing with AWS, and the elegance of Python.

[Igor Gluskin](http://github.com/igorgluskin)

![alt text](https://avatars0.githubusercontent.com/u/22086293?v=3&s=460)

Igor is an undergraduate student at Hunter College, studying Computer Science, with minors in mathematics and economics. He is interested in app development, and other mobile tech projects. He is also a ballroom dance instructor.

[Richard Cibu](http://github.com/r1chc)

Hunter College student--double major in Computer Science and Economics; minor in Political Science. Interested in the fields of Network Security and Big Data Analysis. Currently working at Cravath, Swaine & Moore LLP in the IT department and is an avid soccer player in his spare time.