Remove Static AWS libs. Add link generation.
Static AWS Libraries for the API Gateway are no longer necessary because node.js is handling all of that with the node package "aws-api-gateway-client".

The stylesheet pure-min.css cannot be referenced online because it is served on HTTP. Any source must be served via HTTPS. We included a local copy of the file as a work around.

Added dynamic link generation of any crawl made or retreived. This link is constructed regardless of where the app is hosted.